Fix build error and some checkstyle warnings.

+++ src/main/java/org/spongepowered/mod/mixin/event/block/MixinEventBlock.java
@@ -45,7 +45,7 @@ public abstract class MixinEventBlock extends Event implements BlockEvent {
 
     @Override
     public BlockLoc getBlock() {
-        return new BlockWrapper((World) this.world, (int) this.pos.getX(), (int) this.pos.getY(), (int) this.pos.getZ());
+        return new BlockWrapper((World) this.world, this.pos.getX(), this.pos.getY(), this.pos.getZ());
     }
 
     @Override
